+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added lock package for distributed locks
 * Added memcached cache and lock
 * Added JsonConverter.fromToObject() method
+* Added Fluentd logger
 
 ## <a name="2.7.0"></a> 2.7.0 (2017-09-30)